refactor: index component
