+++ frontend/src/page/AnsPage.tsx
@@ -4,7 +4,7 @@ import ColorSystem from 'utils/ColorSystem';
 import NavBar from 'components/NavBar';
 import Tag from 'components/Tag';
 import TitleIndex from 'components/Index/TitleIndex';
-import ContentIndex from 'components/Index/ContentIndex';
+import ContentIndex from 'components/Index/MiddleIndex';
 import AnswerIndex from 'components/Index/AnswerIndex';
 import Btn from 'components/Btn';
 import Ans from 'components/Ans';
